+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Log message colours can be disabled by setting the `GLEAM_LOG_NOCOLOUR`
   environment variable.
 - You can now specify multiple packages when using `gleam add`
+- If the compiler version changes we now rebuild the project from scratch on
+  next build command (#1547)
 
 ## v0.20.1 - 2022-02-24